requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
djangorestframework-simplejwt==5.3.1
redis==5.0.8
celery==5.4.0
django-celery-beat==2.6.0
python-dotenv==1.0.1
django-filter==24.3
